Adding Stylus storage + stylus repl code with test code
